creates a custom hook for fetching data and refactors single plant page to use the custom hook
